chore: remove duplicate functions in test script
